Refactor coverage analysis
