+++ jdbl-maven-plugin/pom.xml
@@ -156,16 +156,6 @@
          <version>3.8.1</version>
          <scope>test</scope>
       </dependency>
-      <dependency>
-         <groupId>com.fasterxml.jackson.core</groupId>
-         <artifactId>jackson-databind</artifactId>
-         <version>2.9.8</version>
-      </dependency>
-      <dependency>
-         <groupId>se.kth.castor</groupId>
-         <artifactId>yajta-lean</artifactId>
-         <version>2.0.1</version>
-      </dependency>
    </dependencies>
 
 <!--   <profiles>-->
